updates to teaching interface, default to not flush data

+++ testData.py
@@ -10,12 +10,13 @@ from Data.storage import StorageFactory
 from Robot import importer
 
 
-def _flushAndFillTestData(configDir):
+def _flushAndFillTestData(configDir, flush=True):
     StorageFactory.config['engine'].update(dbConfig)
     StorageFactory.config['debug'] = True
-    StorageFactory.drop_keys(StorageFactory.getDefaultDataStore().engine)
-    Base.metadata.drop_all(StorageFactory.getDefaultDataStore().engine)
-    Base.metadata.create_all(StorageFactory.getDefaultDataStore().engine)
+    if flush:
+        StorageFactory.drop_keys(StorageFactory.getDefaultDataStore().engine)
+        Base.metadata.drop_all(StorageFactory.getDefaultDataStore().engine)
+        Base.metadata.create_all(StorageFactory.getDefaultDataStore().engine)
 
     operators = [
                     Operator('Operator1', 'Test Operator 1'),
@@ -68,13 +69,21 @@ def _flushAndFillTestData(configDir):
 if __name__ == "__main__":
     if len(sys.argv) < 2:
         print "Subdir not specified"
-        exit()
-		
+#         exit()
+        subDir = 'kaspar-1c'
+    else:
+        subDir = sys.argv[1]
+
+    if len(sys.argv) > 2:
+        flush = sys.argv[2].lower() == 'flush'
+    else:
+        flush = False
+
     baseDir = os.path.dirname(os.path.realpath(__file__))
-    configDir = os.path.join(baseDir, 'Config/kasparConfigs/' + sys.argv[1])
+    configDir = os.path.join(baseDir, 'Config/kasparConfigs/' + subDir)
 
     if not os.path.isdir(configDir):
-        print "Specified subdir does not exist: %s" % sys.argv[1]
+        print "Specified subdir does not exist: %s" % subDir
         exit()
     else:
-        _flushAndFillTestData(configDir)
+        _flushAndFillTestData(configDir, flush)
